Fix two bug that affect consensus acurracty for the pg_resolve pipeline
The consensus code chopped the the tiling path into ~ 120k chunk now.
There was some conditions where the code missed overlaps at the end
and fall back to the original CCS reads. (we see the errors in the
assembly contigs following some errors in the reads.) We fix this by
add overlaps from the first read in the chunked tiling path reads.

We disable the filtering in filter_ovlp.py accidentally. Add it back,
and change the overlap threshold from 0.96 to 0.99.

We lower the low_cov threhold in shmmer_overlap_annotation.c to
cover low coverage regions better.